+++ src/lobster/model/__init__.py
@@ -5,7 +5,6 @@ from ._clm_configuration import PCLM_CONFIG_ARGS
 from ._conditioanalclassifiermlm import LobsterConditionalClassifierPMLM
 from ._conditioanalmlm import LobsterConditionalPMLM
 from ._dyab import DyAbModel
-from ._esmc import LobsterESMC
 from ._linear_probe import LobsterLinearProbe
 from ._lobster_fold import FoldseekTransform, LobsterPLMFold
 from ._mgm import LobsterMGM
